In that we are Learn Using Spring Data JPA how to perform CRUD operation (using an application context object)

+++ RestBookAPI/src/main/java/com/restBook/RestBookAPI/services/BookServices.java
@@ -20,57 +20,59 @@ public class BookServices {
 //        list.add(new Book(62,"CSS","ABC"));
 //    }
 
+
     @Autowired
     private BookRepository bookRepository;
 
-    //get all book
-    public List<Book> getBookList() {
-        List<Book> allBooks= (List<Book>) bookRepository.findAll();
+    // *************************************************
+    //add Book
+    public Book addBook(Book book)
+    {
+        Book bookSave = bookRepository.save(book);
+        return bookSave;
+    }
+    // *************************************************
+
+    //get allBooks
+    public List<Book> getAllBooks()
+    {
+        List<Book> allBooks = (List<Book>) bookRepository.findAll();
         return allBooks;
     }
+    // *************************************************
 
-    //get single book by id
-    public Book getBookById(int id) {
+    //get a single book by id
+    public Book getBookById(int id)
+    {
         Book book=null;
-        try {
-          //  book = list.stream().filter(e -> e.getId() == id).findFirst().get();
-           book =this.bookRepository.findById(id);
+        try{
+            book = bookRepository.findById(id);
         }
-        catch (Exception e) {
+        catch (Exception e)
+        {
             e.printStackTrace();
         }
-         return book;
+        return book;
     }
-    //adding the book
-    public Book addBook(Book book) {
-        Book result=bookRepository.save(book);
-        return result;
-    }
-    public void deleteBook(int bid) {
-//        list.stream().filter(e->{
-//            if(e.getId()!=bid)
-//            {
-//                  return true;
-//            }
-//            return false;
-//        }).collect(Collectors.toList());
+    // *************************************************
 
-       // list=list.stream().filter(book->book.getId()!=bid).collect(Collectors.toList());
-
+    //delete Book
+    public void deleteBook(int bid)
+    {
+        if(!bookRepository.existsById(bid))
+        {
+            throw new RuntimeException("Book Not Found id:"+bid);
+        }
         bookRepository.deleteById(bid);
     }
 
+    // *************************************************
+
+    //updateBook
     public void updateBook(Book book,int bookId)
     {
-//        list=list.stream().map(b->{
-//            if(b.getId()==bookId)
-//            {
-//                b.setTitle(book.getTitle());
-//                b.setAuthor(book.getAuthor());
-//            }
-//            return b;
-//        }).collect(Collectors.toList());
         book.setId(bookId);
         bookRepository.save(book);
     }
+
 }
